fixing invoice generation

+++ nowellpoint-aws-api/src/main/java/com/nowellpoint/api/invoice/model/AbstractService.java
@@ -1,12 +1,14 @@
 package com.nowellpoint.api.invoice.model;
 
+import java.math.BigDecimal;
+
 import org.immutables.value.Value;
 
 @Value.Immutable
 @Value.Style(typeImmutable = "*", deepImmutablesDetection = true)
 public abstract class AbstractService {
 	public abstract String getServiceName();
-	public abstract Double getUnitPrice();
+	public abstract BigDecimal getUnitPrice();
 	public abstract Integer getQuantity();
-	public abstract Double getTotalPrice();
+	public abstract BigDecimal getTotalPrice();
 }
